Transform the TreeLayoutProcessor into a GEF LayoutManager.
- Create a GraphFreeformLayout from FreeformLayout that can layout as a tree or as XY.
- Make Node.size.height default such that the height is fully manage by the figure (no need to adapt the height based on the font, no need to adapt the height when changing the size of a child figure, no need to initialize the size and location when creating Graph from a product system).
- Remove LayoutCommand execution from MassExpansionCommand as the NodeFigure.s are automatically laid out after any child creation/disposal.
