+++ olca-app/src/org/openlca/app/editors/graphical/edit/GraphXYLayoutEditPolicy.java
@@ -13,7 +13,7 @@ import org.eclipse.gef.requests.ChangeBoundsRequest;
 import org.eclipse.gef.requests.CreateRequest;
 import org.eclipse.osgi.util.NLS;
 import org.openlca.app.M;
-import org.openlca.app.editors.graphical.layouts.TreeLayoutProcessor;
+import org.openlca.app.editors.graphical.layouts.GraphFreeformLayout;
 import org.openlca.app.editors.graphical.model.Node;
 import org.openlca.app.editors.graphical.model.commands.NodeSetConstraintCommand;
 
@@ -54,18 +54,21 @@ public class GraphXYLayoutEditPolicy extends XYLayoutEditPolicy {
 
 	protected Command getLayoutCommand() {
 		var graphEditPart = (GraphEditPart) getHost();
-		var graph = graphEditPart.getModel();
+		var layoutManager = (GraphFreeformLayout) graphEditPart.getFigure()
+			.getLayoutManager();
 
 		CompoundCommand cc = new CompoundCommand();
 		cc.setLabel(NLS.bind(M.LayoutAs.toLowerCase(), M.Tree));
 
-		var layoutProcessor = new TreeLayoutProcessor(graph);
-		var mapNodeToMoveDelta = layoutProcessor.getMoveDeltas();
+		var mapNodeToLocation = layoutManager.layoutProcessor.getNodeLocations();
 
 		for (NodeEditPart part : graphEditPart.getChildren()) {
 			var request = new ChangeBoundsRequest(REQ_MOVE_CHILDREN);
 			request.setEditParts(part);
-			request.setMoveDelta(mapNodeToMoveDelta.get(part.getModel()));
+			var newLoc = mapNodeToLocation.get(part.getFigure());
+			var oldLoc = part.getFigure().getLocation();
+			var moveDelta = newLoc.getTranslated(oldLoc.getNegated());
+			request.setMoveDelta(moveDelta);
 			cc.add(getCommand(request));
 		}
 		return cc.unwrap();
