+++ olca-app/src/org/openlca/app/editors/graphical/figures/NodeFigure.java
@@ -2,6 +2,7 @@ package org.openlca.app.editors.graphical.figures;
 
 import org.eclipse.draw2d.*;
 import org.eclipse.draw2d.geometry.Dimension;
+import org.eclipse.draw2d.geometry.Point;
 import org.eclipse.swt.SWT;
 import org.openlca.app.M;
 import org.openlca.app.editors.graphical.model.Node;
@@ -14,10 +15,11 @@ import static org.openlca.app.editors.graphical.model.Node.Side.OUTPUT;
 
 public class NodeFigure extends Figure {
 
-	final Node node;
+	public final Node node;
 	public final static Dimension HEADER_ARC_SIZE = new Dimension(15, 15);
 	public final PlusMinusButton inputExpandButton;
 	public final PlusMinusButton outputExpandButton;
+	private Point prefLocation;
 
 	public NodeFigure(Node node) {
 		this.node = node;
